backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
numpy
httpx
requests
python-dotenv
faiss-cpu
sentence-transformers
